temp

+++ Assets/UIEffect/UIEffectBase/UIEffectBase.cginc
@@ -13,14 +13,14 @@
 		#define UI_BLUR
 	#endif
 	
-	//把float解析成 fixed4 xyzw 被压缩成0~1
+	//把float解析成 half4 xyzw 被压缩成0~1
 	//float 按照 ((((w)*64+z)*64+y)*64+x) 储存
 	//63为最大精度(6位)
-	fixed4 UnpackToVec4(float value)
+	half4 UnpackToVec4(float value)
 	{
-		const int PACKER_STEP = 64;
-		const int PRECISION = PACKER_STEP - 1;
-		fixed4 unpacked;
+		const uint PACKER_STEP = 64;
+		const uint PRECISION = PACKER_STEP - 1;
+		half4 unpacked;
 		
 		unpacked.x = (value % PACKER_STEP) / PRECISION;
 		value = floor(value / PACKER_STEP);
@@ -39,11 +39,11 @@
 	//把float解析成 fixed3 xyz 被压缩成0~1
 	//float 按照 (((z)*256+y)*256+x) 储存
 	//255为最大精度(8位)
-	fixed3 UnpackToVec3(float value)
+	half3 UnpackToVec3(float value)
 	{
 		const int PACKER_STEP = 256;
 		const int PRECISION = PACKER_STEP - 1;
-		fixed3 unpacked;
+		half3 unpacked;
 		
 		unpacked.x = (value % (PACKER_STEP)) / (PACKER_STEP - 1);
 		value = floor(value / (PACKER_STEP));
